feat: Remove unnecessary enqueueing of `style.css` (#129)
Remove the enqueuing of the `style.css` file. In this block theme, the file is used solely for metadata and is not intended to be a source for front-end styles. And it eliminates one unnecessary HTTP request from the front end.

For more information on the theme's styling architecture, please refer to the updated `README.md` file.